requirements.txt
Django==5.1.5
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
Naked==0.1.32
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-magic>=0.4.24
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
scikit-fuzzy==0.5.0
scipy==1.15.2
service-identity==24.2.0
setuptools==75.8.0
shellescape==3.8.1
simpful==2.12.0
six==1.17.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
zope.interface==7.2
torch
shap

